tailor clang-tidy for Qt project and fix findings
Rewrite .clang-tidy with production-grade checks tailored for Qt 6:
- Exclude Qt false positives (NewDelete, VirtualCall, owning-memory,
  redundant-access-specifiers, etc.)
- Add new checks: cppcoreguidelines-slicing, virtual-class-destructor,
  modernize-loop-convert, use-default-member-init, readability checks
- Remove modernize-use-auto (conflicts with qobject_cast patterns)
- Fix HeaderFilterRegex to exclude libs/KDDockWidgets-2.4.0/src/

Fix the two findings from the new config:
- sensorreadingmodel: widen windowMinutes to qint64 before multiply
- serialhandler: remove redundant m_baudRate constructor initializer

Fix QML type annotation warning (unrelated pre-existing issue):
- Main.qml: use SensorReading type instead of var in onNewReading

+++ src/models/sensorreadingmodel.cpp
@@ -289,7 +289,7 @@ void SensorReadingModel::pruneOldReadings(int windowMinutes)
     if (m_readings.isEmpty())
         return;
 
-    QDateTime cutoff = QDateTime::currentDateTime().addSecs(-windowMinutes * 60);
+    QDateTime cutoff = QDateTime::currentDateTime().addSecs(-static_cast<qint64>(windowMinutes) * 60);
 
     // Find index of first reading to keep
     int firstToKeep = 0;
